Move code to src/ dir (#250)
