+++ src/pack/tgz.go
@@ -5,8 +5,9 @@ import (
 	"path/filepath"
 
 	log "github.com/sirupsen/logrus"
-	"github.com/tarantool/cartridge-cli/common"
-	"github.com/tarantool/cartridge-cli/project"
+
+	"github.com/tarantool/cartridge-cli/src/common"
+	"github.com/tarantool/cartridge-cli/src/project"
 )
 
 func packTgz(projectCtx *project.ProjectCtx) error {
