+++ src/pack/tmp_dir.go
@@ -6,8 +6,9 @@ import (
 	"path/filepath"
 
 	log "github.com/sirupsen/logrus"
-	"github.com/tarantool/cartridge-cli/common"
-	"github.com/tarantool/cartridge-cli/project"
+
+	"github.com/tarantool/cartridge-cli/src/common"
+	"github.com/tarantool/cartridge-cli/src/project"
 )
 
 const (
